ci: add golangci-lint step and config
Pin golangci-lint v1.64.8 via golangci/golangci-lint-action@v6.
Config enables errcheck, govet, ineffassign, staticcheck, unused,
gofmt, goimports, misspell, unconvert and bodyclose, and relaxes
errcheck for tests. Also fix an unnecessary uint8 conversion the
linter flagged in publisher_test.go.

+++ internal/broker/rabbitmq/publisher_test.go
@@ -103,7 +103,7 @@ func TestPublisher_PublishAvatarUploaded_DeliversMessage(t *testing.T) {
 		require.Equal(t, events.RoutingKeyAvatarUploaded, msg.RoutingKey)
 		require.Equal(t, "application/json", msg.ContentType)
 		require.Equal(t, evt.AvatarID, msg.MessageId)
-		require.Equal(t, uint8(amqp.Persistent), msg.DeliveryMode)
+		require.Equal(t, amqp.Persistent, msg.DeliveryMode)
 
 		var got events.AvatarUploadedEvent
 		require.NoError(t, json.Unmarshal(msg.Body, &got))
